server/requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
bcrypt==3.2.0
black==20.8b1
cffi==1.14.3
click==7.1.2
colorama==0.4.4
dnspython==2.0.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT==0.3.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
gunicorn==20.0.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.21.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
PyJWT==1.4.2
pylint==2.6.0
pymongo==3.11.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
